backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
eval-type-backport>=0.2.0
python-dotenv>=1.0,<2.0
PyJWT>=2.8,<3.0
httpx>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
torch>=2.0,<3.0
transformers>=4.36,<5.0
sentencepiece>=0.1.99,<1.0
pyyaml>=6.0,<7.0
reportlab>=4.0,<5.0
arabic-reshaper>=3.0,<4.0
python-bidi>=0.6,<1.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
